Changes the type for steps to i32

+++ src/handler.rs
@@ -29,7 +29,7 @@ impl<'a> ActionHandler<'a> {
         format!("{}\n{}", self.rotator.az, self.rotator.ele)
     }
 
-    pub fn handle_step_test(&mut self, quick_rot_steps: f32) -> Result<String> {
+    pub fn handle_step_test(&mut self, quick_rot_steps: i32) -> Result<String> {
         self.rotator.num_steps = quick_rot_steps;
         info!("Sending {} steps...", quick_rot_steps);
